+++ tests/Fixture/Container/ContainerWithFileBasedAutoloadedEntryPoint-ProxyDefinition.php
@@ -5,6 +5,6 @@ include_once $this->rootDirectory . '/src/Container/Definition/AbstractDefinitio
 include_once $this->rootDirectory . '/tests/Double/TestDefinition.php';
 include_once $this->rootDirectory . '/tests/Double/StubPrototypeDefinition.php';
 
-self::$entryPoints[\WoohooLabs\Zen\Tests\Double\StubPrototypeDefinition::class] = 'WoohooLabs__Zen__Tests__Double__StubPrototypeDefinition';
+self::$entryPoints['WoohooLabs\Zen\Tests\Double\StubPrototypeDefinition'] = 'WoohooLabs__Zen__Tests__Double__StubPrototypeDefinition';
 
 return require __DIR__ . '/WoohooLabs__Zen__Tests__Double__StubPrototypeDefinition.php';