+++ tests/Fixture/Container/ContainerWithOptimizedAutoloadedEntryPoint.php
@@ -9,7 +9,7 @@ class ContainerWithOptimizedAutoloadedEntryPoint extends AbstractCompiledContain
      * @var string[]
      */
     protected static $entryPoints = [
-        \WoohooLabs\Zen\Tests\Double\StubSingletonDefinition::class => 'WoohooLabs__Zen__Tests__Double__StubSingletonDefinition',
+        'WoohooLabs\Zen\Tests\Double\StubSingletonDefinition' => 'WoohooLabs__Zen__Tests__Double__StubSingletonDefinition',
     ];
 
     /**